bu-report-mcp/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.2
cryptography==46.0.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.1.0
markdown-it-py==4.0.0
mcp==1.27.0
mdurl==0.1.2
numpy==2.4.4
pandas==3.0.2
pillow==12.2.0
pycparser==3.0
pydantic==2.13.2
pydantic-settings==2.13.1
pydantic_core==2.46.2
Pygments==2.20.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
python-pptx==1.0.2
referencing==0.37.0
rich==15.0.0
rpds-py==0.30.0
shellingham==1.5.4
six==1.17.0
sse-starlette==3.3.4
starlette==1.0.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
xlsxwriter==3.2.9
